requirements.txt
requests
pandas
openpyxl
matplotlib
streamlit
python-dotenv